"Fix" warnings when compiling for a 64-bit architecture. In most cases, explicitly cast to int. In the Windows version of lou_getProgramPath, change local variable size from type size_t to DWORD because DWORD <= size_t. Change ignoreCaseComp to accept length of type size_t because this corresponds with the signature of the function strncasecmp that it implements.

+++ liblouis/logging.c
@@ -39,7 +39,7 @@ void logWidecharBuf(logLevels level, const char *msg, const widechar *wbuf, int 
    * Give space for additional message (+ strlen(msg))
    * Remember the null terminator (+ 1)
    */
-  int logBufSize = (wlen * ((sizeof(widechar) * 3) + 3)) + 3 + strlen(msg);
+  int logBufSize = (wlen * ((sizeof(widechar) * 3) + 3)) + 3 + (int)strlen(msg);
   char *logMsg = malloc(logBufSize);
   char *p = logMsg;
   char *formatString;
